requirements.txt
python-chess>=1.10.0
psutil>=5.9.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
tqdm>=4.65.0
colorama>=0.4.6
